dictionary: refactor: extract stl_dictionary_wrapper.h as common logic for stl-like dictionaries
